users/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
marshmallow
email-validator
flask-cors
pyjwt
bcrypt
flask-restful
pytest
werkzeug
kafka-python